chore: consolidate geofencing logic for earn calls to action
