+++ app/components/UI/Earn/components/EarnLendingBalance/index.tsx
@@ -43,6 +43,7 @@ import styleSheet from './EarnLendingBalance.styles';
 import { trace, TraceName } from '../../../../../util/trace';
 import { useMusdConversionTokens } from '../../hooks/useMusdConversionTokens';
 import MusdConversionAssetOverviewCta from '../Musd/MusdConversionAssetOverviewCta';
+import { useStakingEligibilityGuard } from '../../../Stake/hooks/useStakingEligibilityGuard';
 
 export const EARN_LENDING_BALANCE_TEST_IDS = {
   RECEIPT_TOKEN_BALANCE_ASSET_LOGO: 'receipt-token-balance-asset-logo',
@@ -82,6 +83,7 @@ const EarnLendingBalance = ({ asset }: EarnLendingBalanceProps) => {
   const { outputToken: receiptToken, earnToken } = useSelector(
     (state: RootState) => selectEarnTokenPair(state, asset),
   );
+  const { checkEligibilityAndRedirect } = useStakingEligibilityGuard();
   const isAssetReceiptToken = useSelector((state: RootState) =>
     selectEarnOutputToken(state, asset),
   );
@@ -141,6 +143,9 @@ const EarnLendingBalance = ({ asset }: EarnLendingBalanceProps) => {
   };
 
   const handleNavigateToWithdrawalInputScreen = async () => {
+    if (!checkEligibilityAndRedirect()) {
+      return;
+    }
     trace({ name: TraceName.EarnWithdrawScreen });
     emitLendingActionButtonMetaMetric('withdrawal');
     const networkClientId = getNetworkClientId(asset);
@@ -159,6 +164,9 @@ const EarnLendingBalance = ({ asset }: EarnLendingBalanceProps) => {
   };
 
   const handleNavigateToDepositInputScreen = async () => {
+    if (!checkEligibilityAndRedirect()) {
+      return;
+    }
     trace({ name: TraceName.EarnDepositScreen });
     emitLendingActionButtonMetaMetric('deposit');
     const networkClientId = getNetworkClientId(asset);
