+++ app/components/UI/Earn/components/Tron/TronStakingButtons/TronStakingButtons.tsx
@@ -20,6 +20,7 @@ import { EVENT_LOCATIONS } from '../../../../../UI/Stake/constants/events';
 import { trace, TraceName } from '../../../../../../util/trace';
 import { RootState } from '../../../../../../reducers';
 import { selectAsset } from '../../../../../../selectors/assets/assets-list';
+import { useStakingEligibilityGuard } from '../../../../../UI/Stake/hooks/useStakingEligibilityGuard';
 
 interface TronStakingButtonsProps extends Pick<ViewProps, 'style'> {
   asset: TokenI;
@@ -38,6 +39,7 @@ const TronStakingButtons = ({
   const { styles } = useStyles(styleSheet, { theme });
   const navigation = useNavigation();
   const { trackEvent, createEventBuilder } = useMetrics();
+  const { checkEligibilityAndRedirect } = useStakingEligibilityGuard();
 
   const isStakedTrx =
     asset?.isStaked || asset?.symbol === 'sTRX' || asset?.ticker === 'sTRX';
@@ -59,6 +61,9 @@ const TronStakingButtons = ({
   );
 
   const onStakePress = () => {
+    if (!checkEligibilityAndRedirect()) {
+      return;
+    }
     trace({ name: TraceName.EarnDepositScreen });
     navigation.navigate('StakeScreens', {
       screen: Routes.STAKING.STAKE,
@@ -76,6 +81,9 @@ const TronStakingButtons = ({
   };
 
   const onUnstakePress = () => {
+    if (!checkEligibilityAndRedirect()) {
+      return;
+    }
     trace({ name: TraceName.EarnWithdrawScreen });
     navigation.navigate('StakeScreens', {
       screen: Routes.STAKING.UNSTAKE,
