+++ app/components/UI/Stake/components/StakingBalance/StakingButtons/StakingButtons.tsx
@@ -19,6 +19,7 @@ import { EVENT_LOCATIONS } from '../../../constants/events';
 import useStakingChain from '../../../hooks/useStakingChain';
 import styleSheet from './StakingButtons.styles';
 import { trace, TraceName } from '../../../../../../util/trace';
+import { useStakingEligibilityGuard } from '../../../hooks/useStakingEligibilityGuard';
 
 interface StakingButtonsProps extends Pick<ViewProps, 'style'> {
   asset: TokenI;
@@ -42,6 +43,7 @@ const StakingButtons = ({
   const isPooledStakingEnabled = useSelector(selectPooledStakingEnabledFlag);
 
   const { isStakingSupportedChain } = useStakingChain();
+  const { checkEligibilityAndRedirect } = useStakingEligibilityGuard();
 
   const { MultichainNetworkController } = Engine.context;
 
@@ -56,6 +58,9 @@ const StakingButtons = ({
   );
 
   const onUnstakePress = async () => {
+    if (!checkEligibilityAndRedirect()) {
+      return;
+    }
     trace({ name: TraceName.EarnWithdrawScreen });
     await handleIsStakingSupportedChain();
     navigate('StakeScreens', {
@@ -77,6 +82,9 @@ const StakingButtons = ({
   };
 
   const onStakePress = async () => {
+    if (!checkEligibilityAndRedirect()) {
+      return;
+    }
     trace({ name: TraceName.EarnDepositScreen });
     await handleIsStakingSupportedChain();
     navigate('StakeScreens', {
